fix(markdown-panel): make its diff split toggle split, and light up correctly
PR #1204 fixed the split-view toggle in SidePanel's Turn Diff tab, but
MarkdownPanel — the file viewer that hosts the same Monaco diff in the
chat side panel and the file explorer pane — owns a second, independent
toggle that kept both of the original bugs. Testing the merged fix on a
file diff therefore still showed no split and an inverted highlight.

Same two causes, same two fixes.

Monaco silently overrode renderSideBySide. monaco-editor 0.55 defaults
useInlineViewWhenSpaceIsLimited to true with a 900px
renderSideBySideInlineBreakpoint, and this panel is well under 900px at
every usable width in both of its hosts, so the persisted default (split)
had never rendered as split. Opt out so renderSideBySide stays
authoritative — the panel owns an explicit toggle, and Monaco must not
second-guess the user's choice on width.

The button's active class was gated on !diffSplit via barIconBtn, whose
argument is "is this control on", so the accent fill marked unified mode
as active — the inverse of the diff-mode button beside it, and
contradicting its own tooltip, which correctly said "Switch to split
view" while the fill claimed split was already on. aria-pressed was
inverted the same way, so assistive tech was told the wrong state too
(WCAG 4.1.2). Both now read the plain state; the tooltip needed no change.

Tests: diffSplitToggles.test.ts (renamed from sidePanelDiffToggle.test.ts,
which only covered SidePanel — nothing pinned MarkdownPanel, which is why
it was missed) now pins both panels' active-class gates, MarkdownPanel's
aria-pressed, and the narrow-width inline opt-out. A class-string
inversion is invisible to tsc, so these assert on source. The Monaco-option
assertion is scoped to the options line rather than the whole file: the
prop's own JSDoc names the same option, so a file-wide search would still
pass after the option was deleted from the editor config. Mutation-checked
— dropping the option and re-inverting the gate each turn the suite red.

Verified against Monaco 0.55.1 at 520px: renderSideBySide alone yields
one 454px pane with the original collapsed to 36px and no side-by-side
class; adding useInlineViewWhenSpaceIsLimited:false yields
monaco-diff-editor.side-by-side with two 245px panes. Screenshots in
temp-screenshots/markdown-panel-split/.
